refactor: change types in .ts files
