+++ src/components/view/news/news.ts
@@ -1,24 +1,26 @@
+
+import { Text } from '../sources/sources';
 import './news.css';
 
-export interface IDataNews {
-    status: string;
-    totalResults: number;
-    articles: Array<Text>;
-};
+// export interface IDataNews {
+//     status: string;
+//     totalResults: number;
+//     articles: Array<Text>;
+// };
 
-export type Text = {
-    source: {
-        id?: string;
-        name: string;
-        };
-    author?: string;
-    title: string;
-    description: string;
-    url: string;
-    urlToImage: string;
-    publishedAt: string;
-    content?: string;
-};
+// export type Text = {
+//     source: {
+//         id?: string;
+//         name: string;
+//         };
+//     author?: string;
+//     title: string;
+//     description: string;
+//     url: string;
+//     urlToImage: string;
+//     publishedAt: string;
+//     content?: string;
+// };
 
 class News {
